74 - more tidy
