+++ src/app/Core/services/user/user.service.ts
@@ -1,35 +1,8 @@
 import { Injectable } from '@angular/core'
-import { FormControl } from '@angular/forms'
 import PocketBase, { ClientResponseError } from 'pocketbase'
 import { ApiService } from 'src/app/Core/services/api/api.service'
+import { User, UserList, UserPassword } from '../../state/user/user'
 import { ErrorService } from '../error/error.service'
-
-export interface User {
-	//ToDo - Read 'https://stackoverflow.com/questions/12789231/class-type-check-in-typescript'
-	id: string
-	username: string
-	email: string | null
-	lastLoggedIn: string | null
-}
-
-export interface UserPassword {
-	username: string
-	email: string
-	emailVisibility: boolean
-	password: string
-	passwordConfirm: string
-}
-
-export interface UserList {
-	id: string
-	username: string
-	email: string | null
-}
-
-export interface UserListSearch {
-	id: FormControl<string | null>
-	username: FormControl<string | null>
-}
 
 @Injectable()
 export class UserService {
